backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
motor==3.3.2
opencv-python-headless==4.9.0.80
mediapipe==0.10.9
numpy==1.26.3
python-dotenv==1.0.1
pymongo==4.6.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic[email]
